02_Brain_Cluster/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
# Forensic tools
cryptography==42.0.0
google-genai>=0.1.0
google-generativeai==0.8.3
PyMuPDF
playwright
Pillow
markdown


# Auth
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
Jinja2==3.1.3

# Download Center
python-docx==1.2.0
beautifulsoup4>=4.12.0
